Feature: Now cell functions can also be written directly.
To do this, the cell value in the input array must begin with an "=".
The function itself must be written in english Excel notation.
Cell values that are to be passed as pure strings and that may happen to begin with an "=" can be written by doubling the "=" at the beginning.